move loadRegion and saveRegion
