+++ server/core/storage.go
@@ -27,6 +27,7 @@ import (
 	"github.com/gogo/protobuf/proto"
 	"github.com/pingcap/errors"
 	"github.com/pingcap/kvproto/pkg/metapb"
+	"github.com/tikv/pd/pkg/encryption"
 	"github.com/tikv/pd/pkg/errs"
 	"github.com/tikv/pd/server/encryptionkm"
 	"github.com/tikv/pd/server/kv"
@@ -576,4 +577,41 @@ func saveProto(s kv.Base, key string, msg proto.Message) error {
 		return errs.ErrProtoMarshal.Wrap(err).GenWithStackByCause()
 	}
 	return s.Save(key, string(value))
+}
+
+func loadRegion(
+	kv kv.Base,
+	encryptionKeyManager *encryptionkm.KeyManager,
+	regionID uint64,
+	region *metapb.Region,
+) (ok bool, err error) {
+	value, err := kv.Load(regionPath(regionID))
+	if err != nil {
+		return false, err
+	}
+	if value == "" {
+		return false, nil
+	}
+	err = proto.Unmarshal([]byte(value), region)
+	if err != nil {
+		return true, errs.ErrProtoUnmarshal.Wrap(err).GenWithStackByArgs()
+	}
+	err = encryption.DecryptRegion(region, encryptionKeyManager)
+	return true, err
+}
+
+func saveRegion(
+	kv kv.Base,
+	encryptionKeyManager *encryptionkm.KeyManager,
+	region *metapb.Region,
+) error {
+	err := encryption.EncryptRegion(region, encryptionKeyManager)
+	if err != nil {
+		return err
+	}
+	value, err := proto.Marshal(region)
+	if err != nil {
+		return errs.ErrProtoMarshal.Wrap(err).GenWithStackByArgs()
+	}
+	return kv.Save(regionPath(region.GetId()), string(value))
 }
